Corrected one type mismatch, accepted 4.5 updates. I tried loading one complicated Collada file, and crashed with "index overflow" after warning "Mesh has neither <triangles> nor <polylist>". Did not investigate at all.

+++ iOS examples/ch07_EXAMPLES/COLLADAViewer/COLLADAViewer/COLLADAGeometryInfo.m
@@ -31,7 +31,7 @@
    *mesh;
 @property (nonatomic, readwrite, strong) NSMutableSet 
    *existingCombinations;
-@property (nonatomic, readwrite, assign) NSUInteger 
+@property (nonatomic, readwrite, assign) GLushort 
    nextAvalableIndex;
 @property (retain, nonatomic, readwrite) NSMutableDictionary 
    *mutableSourcesByID;
